% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_encode_data}
\alias{clipboard_encode_data}
\title{Prepare Data for Clipboard Copy}
\usage{
clipboard_encode_data(data, format_name)
}
\arguments{
\item{data}{data to be prepared}

\item{format_name}{Clipboard image format to be prepared.}
}
\value{
Converted data, as raw vector.
}
\description{
Converts data and prepares it for the clipboard.
}
