% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_encode_text}
\alias{clipboard_encode_text}
\title{Convert data to raw}
\usage{
clipboard_encode_text(data, format_name)
}
\arguments{
\item{format_name}{Clipboard image format to be encoded.}
}
\value{
raw data
}
\description{
Internal function that converts data to raw.
Similar to \code{charToRaw()}
}
\details{
We need to do modifications depending on the format.
For \code{CF_TEXT} and \code{CF_UNICODETEXT}, we need to add bytes at the end,
And for \code{CF_UNICODETEXT} we need UTF-16LE.
}
