% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_inspect_dib}
\alias{clipboard_inspect_dib}
\title{Inspect DIB Image on Clipboard}
\usage{
clipboard_inspect_dib()
}
\value{
A list containing DIB image metadata.
}
\description{
Inspect DIB Image on Clipboard
}
