% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clipboard_inspect_windows}
\alias{clipboard_inspect_windows}
\title{Inspect Clipboard on Windows}
\usage{
clipboard_inspect_windows()
}
\value{
data frame showing the various formats available on the clipboard
}
\description{
Inspect Clipboard on Windows
}
