% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_isolate}
\alias{clipboard_isolate}
\title{Isolate Clipboard Elements}
\usage{
clipboard_isolate(...)
}
\arguments{
\item{...}{Formats to be isolated. Can be ID (integer) or name (character).
These are listed with \code{\link[=clipboard_inspect]{clipboard_inspect()}}}
}
\description{
Isolate Clipboard Elements
}
