% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_read_dib}
\alias{clipboard_read_dib}
\title{Read DIB Image Data from Clipboard}
\usage{
clipboard_read_dib()
}
\value{
A raw vector containing Windows DIB data.
}
\description{
Read DIB Image Data from Clipboard
}
