% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_read_image}
\alias{clipboard_read_image}
\title{Read an Image from the Windows Clipboard}
\usage{
clipboard_read_image(format_name = NULL)
}
\arguments{
\item{format_name}{Clipboard image format to read. If \code{NULL}, the first
available format from the preferred image-format order is used.}
}
\value{
A \code{magick-image} object.
}
\description{
Read an Image from the Windows Clipboard
}
