% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_read_text}
\alias{clipboard_read_text}
\title{Read Dat from Clipboard}
\usage{
clipboard_read_text(format_name)
}
\arguments{
\item{format_name}{the desired clipboard format. E.g., \code{"Rich Text Format"}}
}
\description{
Read Dat from Clipboard
}
