% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_write_raw}
\alias{clipboard_write_raw}
\title{Write Raw Text to Clipboard}
\usage{
clipboard_write_raw(data, format_name)
}
\arguments{
\item{data}{Data to write.}

\item{format_name}{Clipboard format.}
}
\description{
Write Raw Text to Clipboard
}
