% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkprojects.R
\name{dplyr_or_base_status}
\alias{dplyr_or_base_status}
\title{Turn the raw signals into a single human-readable status label}
\usage{
dplyr_or_base_status(is_dirty, ahead, behind, has_upstream)
}
\description{
Turn the raw signals into a single human-readable status label
}
