% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkprojects.R
\name{find_git_repos}
\alias{find_git_repos}
\title{Find all directories containing a .git folder under given root(s)}
\usage{
find_git_repos(roots, max_depth = 5)
}
\description{
Find all directories containing a .git folder under given root(s)
}
