% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkprojects.R
\name{get_repo_status}
\alias{get_repo_status}
\title{Get the status of a single repo as a one-row data frame}
\usage{
get_repo_status(repo_path)
}
\description{
Get the status of a single repo as a one-row data frame
}
