% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkprojects.R
\name{git_cmd}
\alias{git_cmd}
\title{Run a git command inside a given repo directory and return trimmed stdout}
\usage{
git_cmd(repo_path, args)
}
\description{
Run a git command inside a given repo directory and return trimmed stdout
}
