% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{knit_code_OLD}
\alias{knit_code_OLD}
\title{Convert R Code to Formatted HTML}
\usage{
knit_code_OLD(code, options = list())
}
\arguments{
\item{options}{}
}
\description{
Convert R Code to Formatted HTML
}
