% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{magick_to_cf_dib}
\alias{magick_to_cf_dib}
\title{Convert a magick image to an Office-compatible CF_DIB}
\usage{
magick_to_cf_dib(img)
}
\arguments{
\item{img}{A single \code{magick-image} object.}
}
\value{
A raw vector containing the \code{CF_DIB} data, suitable for use with
\code{\link[=clipboard_write_formats]{clipboard_write_formats()}}.
}
\description{
Converts a single \code{magick-image} object to the raw representation of a
Windows \code{CF_DIB} clipboard format. The DIB uses a 32-bit
\code{BITMAPINFOHEADER} with \code{BI_BITFIELDS} RGB masks and bottom-up BGR0 pixel
data. This representation is compatible with Microsoft Office
applications such as Word and PowerPoint.
}
\details{
This is an internal helper used by \code{\link[=clipboard_write_image]{clipboard_write_image()}} and is not
intended to be called directly by users.
}
\keyword{internal}
