% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_code.R
\name{read.clipboard.rterminal}
\alias{read.clipboard.rterminal}
\title{Convert R Terminal Output to Rich Text}
\usage{
read.clipboard.rterminal(background = NULL)
}
\description{
Convert R Terminal Output to Rich Text
}
