% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_code.R
\name{rstudio_html_restyle}
\alias{rstudio_html_restyle}
\title{Restyle RStudio-generated HTML using another theme}
\usage{
rstudio_html_restyle(html, theme)
}
\arguments{
\item{html}{A single character string containing RStudio-generated
HTML.}

\item{theme}{An RStudio theme object returned by
\code{\link[=read_rstudio_theme]{read_rstudio_theme()}}.}
}
\value{
A character string containing the restyled HTML.
}
\description{
Translates RStudio console HTML to the colors defined by a
destination RStudio theme. The HTML is expected to have been
generated by RStudio and to contain RStudio terminal and
syntax-highlighting classes.
}
\details{
The console foreground and background colors are obtained from the
destination theme's \code{.terminal} rule. Syntax-highlighted text is
restyled using the CSS class rules in the destination theme,
including \code{xtermColor} and Ace editor classes.

The source theme does not need to be specified because the HTML
contains the classes needed to identify the corresponding
destination colors.
}
\examples{
\dontrun{
html <- yclip::clipboard_read_text("HTML Format")
dracula <- read_rstudio_theme("dracula.rstheme")
html_dracula <- rstudio_html_restyle(html, dracula)
}
}
