% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_code.R
\name{view_html}
\alias{view_html}
\title{View HTML in Viewer}
\usage{
view_html(html, title = "yclip HTML")
}
\arguments{
\item{html}{character vector with HTML}

\item{title}{}
}
\description{
View HTML in Viewer
}
