% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{word_convert_rtf_to_html}
\alias{word_convert_rtf_to_html}
\title{Convert RTF text to HTML}
\usage{
word_convert_rtf_to_html(rtf)
}
\arguments{
\item{rtf}{character RTF data}
}
\value{
character HTML data
}
\description{
Convert RTF text to HTML
}
