% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_read_raw}
\alias{clipboard_read_raw}
\title{Read Raw Data from Clipboard}
\usage{
clipboard_read_raw(format_name)
}
\arguments{
\item{format_name}{the desired clipboard format. E.g., \code{"Rich Text Format"}}
}
\description{
Read Raw Data from Clipboard
}
