% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{clipboard_write_formats}
\alias{clipboard_write_formats}
\title{Write Multiple Formats to the Windows Clipboard}
\usage{
clipboard_write_formats(...)
}
\arguments{
\item{...}{Named raw vectors, with the names being clipboard
format names.}
}
\value{
Invisibly returns NULL.
}
\description{
Write Multiple Formats to the Windows Clipboard
}
