% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_code.R
\name{read.clipboard.rcode}
\alias{read.clipboard.rcode}
\title{Convert R Code on Clipboard to Rich Text With Syntax Highlighting}
\usage{
read.clipboard.rcode()
}
\description{
Convert R Code on Clipboard to Rich Text With Syntax Highlighting
}
