% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yclip-package.R
\docType{package}
\name{yclip-package}
\alias{yclip}
\alias{yclip-package}
\title{yclip: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
